added game

+++ calc.py
@@ -1,8 +1,8 @@
 name1 =input ("input your first name:")
 print("Hello " +name1)
-num1 = float(input("Enter your first number:"))
+num1 = int(input("Enter your first number:"))
 op = input("Enter your operator:")
-num3 = float(input("Enter your second number:"))
+num3 = int(input("Enter your second number:"))
 if op == '*':
     print(num1 * num3)
 elif op == '-':
